aplay - Fix a compile warning
aplay.c: In function ‘compute_max_peak’:
aplay.c:1327: warning: format ‘%d’ expects type ‘int’, but argument 3 has type ‘size_t’

+++ aplay/aplay.c
@@ -1324,7 +1324,7 @@ static void compute_max_peak(u_char *data, size_t count)
 	}
 	default:
 		if (run == 0) {
-			fprintf(stderr, _("Unsupported bit size %d.\n"), bits_per_sample);
+			fprintf(stderr, _("Unsupported bit size %d.\n"), (int)bits_per_sample);
 			run = 1;
 		}
 		return;
